requirements.txt
tqdm
ollama
langchain
langchain-ollama
langchain-google-genai
python-dotenv
prompt_toolkit
streamlit
google-cloud-firestore
google-auth
streamlit-cookies-controller
